refactor: move oversized inline test blocks to sibling tests.rs files (#2049)
Rule 03 requires non-trivial unit tests to live in a sibling
<mod>/tests.rs rather than an inline #[cfg(test)] mod tests block.
Moves the 15 largest offenders (frontend components/pages/rpc, and
shared/src/worker.rs) to that shape. Pure file reorganization: no test
logic or behavior changes. Progress on #2049 — 37 more inline blocks
in the 50-91 line range remain for follow-up PRs.
